+++ tools/convert_weights/__main__.py
@@ -21,6 +21,7 @@ import tools.convert_weights.convnext  # noqa: F401
 import tools.convert_weights.crossvit  # noqa: F401
 import tools.convert_weights.cspnet  # noqa: F401
 import tools.convert_weights.cvt  # noqa: F401
+import tools.convert_weights.ddpm  # noqa: F401
 import tools.convert_weights.densenet  # noqa: F401
 import tools.convert_weights.detr  # noqa: F401
 import tools.convert_weights.efficientformer  # noqa: F401